Refactor inline AudioWorklet into separate file
Move `CaptureProcessor` from inline string in `AudioEngine.ts` to `src/lib/audio/capture-processor.ts`.
This improves code readability, maintainability, and enables type checking.
The processor is now imported using Vite's `?worker&url` suffix and loaded via `addModule`.

- Extracted `CaptureProcessor` logic to `src/lib/audio/capture-processor.ts`
- Added TypeScript types and constants to the processor
- Updated `AudioEngine.ts` to import the processor URL
- Verified with build and instantiation tests
